requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
pytz==2024.1
PyYAML==6.0.2
setuptools==74.1.2
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
